création/modification/suppression si utilisateur

+++ app/controllers/gossips_controller.rb
@@ -1,4 +1,6 @@
 class GossipsController < ApplicationController
+  before_action :authenticate_user, only: [:new, :create, :edit, :update, :destroy]
+
   def new
     @gossip = Gossip.new
   end
@@ -6,7 +8,7 @@ class GossipsController < ApplicationController
   # possibilité de retoruver l'attribut flash dans application.html.erb
   # le bootstrap des bannières peut être retrouvé dans helper > application helper
   def create
-    @gossip = Gossip.new(title: params[:title], content: params[:content],  user: User.find(params[:user]))
+    @gossip = Gossip.new(title: params[:title], content: params[:content],  user: current_user)
     if @gossip.save
       flash[:success] = "The super potin was succesfully saved !"
       redirect_to gossips_path
@@ -28,36 +30,58 @@ class GossipsController < ApplicationController
   end
 
   def index
-    @gossips = Gossip.all
+      @gossips = Gossip.all
   end 
 
   def edit
-    @gossip = Gossip.find(params[:id])
+      @gossip = Gossip.find(params[:id]) 
+      if !current_user?(@gossip.user)
+        flash[:error] = "Vous n'êtes pas le bon utilisateur."
+        redirect_to gossips_path
+     end
   end
 
   def update
-    @gossip = Gossip.find(params[:id])
-    if @gossip.update(gossip_params)
-      flash[:success] = "Ton super potin a correctement été mis à jour !"
-      redirect_to gossip_path(@gossip.id)
-    else
-      render 'edit'
-    end
+      @gossip = Gossip.find(params[:id])
+      if !current_user?(@gossip.user)
+        flash[:error] = "Vous n'êtes pas le bon utilisateur."
+        redirect_to gossips_path
+      else
+        if @gossip.update(gossip_params)
+          flash[:success] = "Ton super potin a correctement été mis à jour !"
+          redirect_to gossip_path(@gossip.id)
+        else
+          render 'edit'
+        end
+      end
   end
 
   def destroy
-    @gossip = Gossip.find(params[:id])
-    if @gossip.destroy
-      flash[:success] = "Ton super potin a correctement été détruit !"
-      redirect_to gossips_path
-    else
-      render'destroy'
+      @gossip = Gossip.find(params[:id])
+      if !current_user?(@gossip.user)
+        flash[:error] = "Vous n'êtes pas le bon utilisateur."
+        redirect_to gossips_path
+      else
+        if @gossip.destroy
+          flash[:success] = "Ton super potin a correctement été détruit !"
+          redirect_to gossips_path
+        else
+          render'destroy'
+        end
+      end
+  end
+
+  private
+
+  def gossip_params
+    params.require(:gossip).permit(:title, :content)
+  end
+  
+  def authenticate_user
+    unless current_user
+      flash[:error] = "Vous devez vous connecter pour accéder au contenu."
+      redirect_to new_session_path
     end
   end
-end
 
-private
-
-def gossip_params
-  params.require(:gossip).permit(:title, :content)
 end
